+++ service/task.go
@@ -7,8 +7,8 @@ import (
 	"github.com/airportr/miaospeed/utils"
 )
 
-var SpeedTaskPoll *taskpoll.TaskPollController
-var ConnTaskPoll *taskpoll.TaskPollController
+var SpeedTaskPoll *taskpoll.TPController
+var ConnTaskPoll *taskpoll.TPController
 
 func StartTaskServer() {
 	SpeedTaskPoll = taskpoll.NewTaskPollController("SpeedPoll", 1, time.Duration(utils.GCFG.PauseSecond)*time.Second, 200*time.Millisecond)