+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-miaospeed v4.4.9
+miaospeed v4.5.0
 
-1. 建立连接出错时，打印更多的错误信息。
+1. 将一些任务队列里的变量改成原子变量
+2. 提供不是并发安全的用来获取当前任务队列大概长度的方法
+3. 更详细的显示日志信息
+4. 新的Matrix: TEST_HTTP_CODE HTTPing的响应状态码
+5. 新的Matrix: TEST_PING_TOTAL_RTT 本次节点测试的所有rtt数值
+6. 新的Matrix: TEST_PING_TOTAL_CONN 本次节点测试的所有http延迟数值